Refactor logging mechanism to use 'nexoraldns-shared' logger
- Replaced direct imports of the logger utility from local utilities with imports from 'nexoraldns-shared' across multiple service and controller files.
- Updated test files to mock the logger from 'nexoraldns-shared' instead of the local logger utility.
- Removed the local logger utility file and its associated tests.
- Ensured consistent logging behavior across the application by centralizing the logger configuration.
