+++ Test/web/database/MongoCollectionManager.test.ts
@@ -2,9 +2,10 @@ import { describe, it, expect, vi } from 'vitest';
 import { MongoCollectionManager } from '@web/Database/MongoCollectionManager';
 import type { MongoConnectionManager } from '@nexoralShared/Database/MongoConnectionManager';
 
-vi.mock('@web/utilities/logger', () => ({
-  default: { info: vi.fn(), error: vi.fn(), warn: vi.fn() },
-}));
+vi.mock('nexoraldns-shared', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('nexoraldns-shared')>();
+  return { ...actual, logger: { info: vi.fn(), error: vi.fn(), warn: vi.fn() } };
+});
 
 function createFakeConnectionManager() {
   const collectionCalls: string[] = [];
