+++ Test/web/services/DNS/DNS_DoT.Service.test.ts
@@ -28,7 +28,10 @@ vi.mock('node:fs', () => ({
 }));
 vi.mock('node:child_process', () => ({ execFileSync: execFileSyncMock }));
 vi.mock('@web/container/appContainer', () => ({ default: mockContainer }));
-vi.mock('@web/utilities/logger', () => ({ default: { info: vi.fn(), error: vi.fn(), warn: vi.fn() } }));
+vi.mock('nexoraldns-shared', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('nexoraldns-shared')>();
+  return { ...actual, logger: { info: vi.fn(), error: vi.fn(), warn: vi.fn() } };
+});
 vi.mock('@web/utilities/GetWLANIP.utls', () => ({ default: getLocalIPMock }));
 
 function lenPrefixed(payload: Buffer): Buffer {
