+++ Test/web/services/Forwarder/GlobalDNSforwarder.service.test.ts
@@ -11,9 +11,10 @@ const { mockContainer, createSocketMock } = vi.hoisted(() => ({
 
 vi.mock('node:dgram', () => ({ default: { createSocket: createSocketMock } }));
 vi.mock('@web/container/appContainer', () => ({ default: mockContainer }));
-vi.mock('@web/utilities/logger', () => ({
-  default: { info: vi.fn(), error: vi.fn(), warn: vi.fn() },
-}));
+vi.mock('nexoraldns-shared', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('nexoraldns-shared')>();
+  return { ...actual, logger: { info: vi.fn(), error: vi.fn(), warn: vi.fn() } };
+});
 
 const RINFO: dgram.RemoteInfo = { address: '10.0.0.42', port: 5353, family: 'IPv4', size: 0 };
