+++ Web/src/Database/MongoCollectionManager.ts
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Collection, Document } from 'mongodb';
-import logger from '../utilities/logger';
-import { MongoConnectionManager } from 'nexoraldns-shared';
+import { logger, MongoConnectionManager } from 'nexoraldns-shared';
 import { DB_DEFAULT_CONFIGS } from '../Config/key';
 
 export class MongoCollectionManager {
