+++ Web/src/cluster/Cluster.ts
@@ -2,7 +2,7 @@
 import cluster from "cluster";
 import os from "os";
 import handler from "../Config/DNS";
-import logger from "../utilities/logger";
+import { logger } from 'nexoraldns-shared';
 import { loadOrGenerateCerts } from "../services/DNS/DNS_DoT.Service";
 const numCPUs = os.cpus().length;
 const totalUsableCpus = Math.max(1, Math.floor(numCPUs * 0.75)); // Use at least 1 CPU, up to 75% of total CPUs
