+++ Web/src/services/Forwarder/GlobalDNSforwarder.service.ts
@@ -1,8 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import dgram from "node:dgram";
-import logger from "../../utilities/logger"
 
-import { CacheKeys, DNS_QUERY_STATUS_KEYS, QueueKeys, RabbitMQService } from "nexoraldns-shared";
+import { logger, CacheKeys, DNS_QUERY_STATUS_KEYS, QueueKeys, RabbitMQService } from 'nexoraldns-shared';
 import container from "../../container/appContainer";
 import { RedisCacheService } from "../../Redis/Redis.cache";
 import InputOutputHandler from "../../utilities/IO.utls";
