+++ server/source/Services/DNS/Add_DNS.service.ts
@@ -1,4 +1,3 @@
-import logger from '../../utilities/logger';
 import container from '../../container/appContainer';
 import { MongoCollectionManager } from '../../Database/MongoCollectionManager';
 import { FastifyReply } from "fastify";
@@ -9,7 +8,7 @@ import BuildResponse from "../../helper/responseBuilder.helper";
 // keys import
 import { DB_DEFAULT_CONFIGS } from "../../core/key";
 import { RedisCacheService } from "../../Redis/Redis.cache";
-import { CacheKeys } from "nexoraldns-shared";
+import { logger, CacheKeys } from 'nexoraldns-shared';
 
 // db connections
 import { ObjectId } from "mongodb";
